Crate bdk_bitcoind_rpc

This crate is used for emitting blockchain data from the bitcoind RPC interface. It does not use the wallet RPC API, so this crate can be used with wallet-disabled Bitcoin Core nodes.

Emitter is the main structure which sources blockchain data from [bitcoincore_rpc::Client].

To only get block updates (exclude mempool transactions), the caller can use Emitter::next_block or/and Emitter::next_header until it returns Ok(None) (which means the chain tip is reached). A separate method, Emitter::mempool can be used to emit the whole mempool.

Re-exports

pub use bitcoincore_rpc;

Structs

BlockEvent

A newly emitted block from Emitter.

Emitter

The Emitter is used to emit data sourced from [bitcoincore_rpc::Client].

Traits

BitcoindRpcErrorExt

Extends [bitcoincore_rpc::Error].