bdk_bitcoind_rpc

Module bip158

Compact block filters sync over RPC. For more details refer to BIP157.

This module is home to FilterIter, a structure that returns bitcoin blocks by matching a list of script pubkeys against a BIP158 [BlockFilter].

Structs

EventInner

Event inner type

FilterIter

Type that generates block Events by matching a list of script pubkeys against a [BlockFilter].

Enums

Error

Errors that may occur during a compact filters sync.

Event

Kind of event produced by FilterIter.